+++ tests/poll-observable-inband.c
@@ -52,7 +52,7 @@ static void *observable_poller(void *arg)
 		__Tcall_assert(ret, evl_read_observable(ofd, &nf, 1));
 		__Texpr_assert(ret == 1);
 		__Texpr_assert(nf.tag == EVL_NOTICE_USER);
-		__Texpr_assert(nf.event.val == 0xa5a5a5a5);
+		__Texpr_assert(nf.event.val == (int)0xa5a5a5a5);
 	}
 
 	return NULL;
